done day 1

+++ twitoff/models.py
@@ -5,7 +5,7 @@ DB = SQLAlchemy()
 
 class User(DB.Model):
     """Twitter users that we pull and analyze Tweets for."""
-    id = DB.Column(DB.Integer,primary_key=True)
+    id = DB.Column(DB.BigInteger,primary_key=True)
     name = DB.Column(DB.String(15), nullable=False)
 
     def __repr__(self):
@@ -13,10 +13,10 @@ class User(DB.Model):
 
 class Tweet(DB.Model):
     """Tweets."""
-    id = DB.Column(DB.Integer,primary_key=True)
+    id = DB.Column(DB.BigInteger,primary_key=True)
     text = DB.Column(DB.Unicode(280))
-    user_id = DB.Column(DB.Integer, DB.ForeignKey('user.id'), nullable=False)
-    user = DB.relationship('User', backref=DB.backref('tweets', lazy = True))
+    user_id = DB.Column(DB.BigInteger, DB.ForeignKey('user.id'), nullable=False)
+    user = DB.relationship('User', backref=DB.backref('tweets', lazy=True))
 
     def __repr__(self):
         return '<Tweet {}>'.format(self.text)
